requirements.txt
asgiref==3.7.2
Django==5.0.3
mysqlclient==2.2.4
pillow==10.2.0
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.10.0
